refactor(eval): split scheduled learning components
